requirements.txt
requests==2.25.1
pandas==1.3.0
google-cloud-bigquery==2.31.0
google-auth==1.29.0
argparse==1.4.0
pyarrow==6.0.1